moved test to test_proxy_server
